requirements.txt
# Core async web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# WebSocket client for market data
websockets>=12.0

# Telegram bot (async version)
python-telegram-bot>=20.7

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Technical analysis indicators
ta>=0.10.2

# Async SQLite database
aiosqlite>=0.19.0

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0
